[Refactor]추천 AI서비스 분리 - repository 오류 수정
