+++ src/main/java/com/codeit/weatherwear/domain/clothes/repository/ClothRepository.java
@@ -27,7 +27,7 @@ public interface ClothRepository extends JpaRepository<Cloth, UUID>, ClothReposi
   List<Cloth> findAllWithAttributesByUserId(UUID userId);
 
   @EntityGraph(attributePaths = {"clothesWithAttributes", "clothesWithAttributes.attribute"})
-  @Query("SELECT c FROM Cloth c WHERE c.name = :names")
+  @Query("SELECT c FROM Cloth c WHERE c.name IN :names")
   List<Cloth> findAllByNames(List<String> names);
 
 }
