+++ src/main/java/com/codeit/weatherwear/domain/clothes/service/ClothServiceImpl.java
@@ -80,7 +80,7 @@ public class ClothServiceImpl implements ClothService {
 
     //이름 중복 방지
     if (clothRepository.existsByName(request.name())) {
-      throw new IllegalArgumentException("이미 존재하는 이름의 의상입니다.");
+      throw new RuntimeException("이미 존재하는 이름읍니다.");
     }
 
     // 썸네일 S3 업로드
